Archivo backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
openai==1.35.13 # Versión específica, asegúrate de usar una compatible
# Para la base de datos (descomenta y configura según tu elección)
# psycopg2-binary==2.9.9 # Para PostgreSQL
# pymongo==4.7.3 # Para MongoDB 